cleanup and optimize web worker
