+++ src/App.jsx
@@ -3,7 +3,7 @@ import { Game } from './Game';
 import { BoardWrapper } from './components/BoardWrapper';
 import { parseUrl } from './Parsing';
 import { Footer } from './components/Footer';
-import { useEffect, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 
 const wrapBoardWithReload = ({ reload, board: RawBoard }) => {
 	const Board = (props) => {
@@ -16,36 +16,53 @@ const wrapBoardWithReload = ({ reload, board: RawBoard }) => {
 export const App = () => {
 	const { seed, setupData } = parseUrl();
 	const [game, setGame] = useState(null)
+	const [worker, setWorker] = useState();
 
-	const reload = () => {
-		setGame(null);
+	const setupGame = () => {
+		console.log(`Loading game: ${setupData.gamemode === 'c' ? "classic" : "puzzle"} gamemode${seed != null ? ` with a seed of \"${seed}\"` : ""}, ${setupData.count} piece${setupData.count > 1 ? "s" : ""}, ${setupData.size}x${setupData.size} grid, piece${Object.keys(setupData.pieces).length > 1 ? "s" : ""} allowed: ${Object.keys(setupData.pieces).map(x => `${x} (x${setupData.pieces[x]})`).join(', ')}`)
+
+		if (setupData.gamemode === 'p') {
+			worker.postMessage(setupData);
+		} else {
+			setGame(Game(setupData));
+		}
 	}
 
+	// Setup web worker
 	useEffect(() => {
-		let w;
-		if (game === null) {
-			w = new Worker("src/PuzzleGenWebWorker.js", { type: 'module' });
-			w.postMessage(setupData);
-			w.onmessage = (event) => {
-				// todo, handle errors
-				const { cells, knownCells } = event.data;
-				setGame({...Game({...setupData, cells, knownCells}), seed});
-			};	
-		}
+		const w = new Worker("src/PuzzleGenWebWorker.js", { type: 'module' });
+		w.onmessage = (e) => {
+			if (typeof e.data === 'string') {
+				console.error(e.data);
+			} else {
+				const { cells, knownCells } = e.data;
+				setGame(Game({...setupData, cells, knownCells}));
+			}
+		};	
 
-		return () => { if (w) { w.terminate(); }};
-	}, [game, setGame]);
+		setWorker(w);
+
+		return () => {
+			w.terminate();
+			setWorker(undefined);
+		}
+	}, [setGame]);
+
+	// Wait for worker creation to start game
+	useEffect(() => {
+		if (worker) {
+			setupGame();
+		}
+	}, [worker]);
 
 	const Client = game ? BgioClient({
 		game,
-		board: wrapBoardWithReload({ reload, board: BoardWrapper }),
+		board: wrapBoardWithReload({ reload: setupGame, board: BoardWrapper }),
 		numPlayers: 1,
 		debug: {
 			collapseOnLoad: true,
 		}
 	}) : () => <div>Generating Board...</div>;
-
-	console.log(`Loading game: ${setupData.gamemode === 'c' ? "classic" : "puzzle"} gamemode${seed != null ? ` with a seed of \"${seed}\"` : ""}, ${setupData.count} piece${setupData.count > 1 ? "s" : ""}, ${setupData.size}x${setupData.size} grid, piece${Object.keys(setupData.pieces).length > 1 ? "s" : ""} allowed: ${Object.keys(setupData.pieces).map(x => `${x} (x${setupData.pieces[x]})`).join(', ')}`)
 	
 	return (
 		<div>
